refactor(compare): extract C11 to checks/type_consistency.py (Phase A 8/N)
Move check_c11_type_consistency into the checks subpackage. C11 imports
_codes_are_numeric_or_sentinel and _distribution_count_map from crosswalk
(those are the only crosswalk-private helpers it needs).

Regression-check: COPDGene, ARIC, HCHS — MD + JSON all match. Tests:
79/79 pass.

Main file: 1274 → 1237 lines.
